[UPDATE] rabbit 2.0

+++ src/main/java/com/example/productmicroservicio/web/dtos/requests/CreateBuyRequest.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class CreateBuyRequest {
     private Long productId;
-    private Double amount;
+    private Long amount;
     private String sessionId;
 }
